22.06.21 Bug fixing of IsotopePatternLogics, test_IsotopePatternLogics.py unfinished

+++ src/entities/Ions.py
@@ -75,7 +75,7 @@ class Fragment(object):
     def getNumberOfHighestIsotopes(self):
         '''
         Defines number of peaks to be searched for in first step of findIons function in SpectrumHandler
-        :return: number of peaks
+        :return: (int) number of peaks
         '''
         abundances = self._isotopePattern['calcInt'] / self._isotopePattern[0]['calcInt']
         if len(abundances) < 3:
